refactor: CR 피드백 반영 (컴포넌트/테스트 경로 마이그레이션 및 즉각 렌더링 적용)
